Extract Auth and Rate-Limit Middleware from the HTTP Handler (queue-ayv.3)
